+++ tests/test_predicates.py
@@ -11,16 +11,6 @@ from .fixtures import parametrize
 
 
 @parametrize('case,expected', [
-    (({'a': 1, 'b': 2, 'c': 3}, 'b'), True),
-    (([1, 2, 3], 0), True),
-    (([1, 2, 3], 1), True),
-    (([1, 2, 3], 3), False),
-])
-def test_has(case, expected):
-    assert pyd.has(*case) == expected
-
-
-@parametrize('case,expected', [
     ([], True),
     ({}, True),
     ('', True),