+++ CHANGES.rst
@@ -18,6 +18,8 @@ v2.2.0-dev (xxxx-xx-xx)
 - Add ``unshift``.
 - Add ``url``.
 - Fix bug in ``snake_case`` that resulted in returned string not being converted to lower case.
+- Make ``has`` work with deep paths.
+- Make ``has_path`` and ``deep_has`` aliases of ``has``.
 - Make ``get_path`` handle escaping the ``.`` delimiter for string keys.
 - Make ``concat`` and alias of ``cat``.